Patin_Clement_3_application_Flask_042024/requirements.txt
Flask==3.0.3
numpy==1.26.4
Pillow==10.3.0
Requests==2.31.0
python-multipart==0.0.9